move files to numpy_files
